Refactor the release tool (#22003)
